requirements.txt
ccxt==4.2.0
pandas==2.2.0
numpy==1.26.0
matplotlib==3.8.0
tk==8.6.12
python-dotenv==1.0.0
tenacity==8.2.3